noms_images

+++ Back/controllers/sauce.js
@@ -5,9 +5,9 @@ const fs = require("fs");
 exports.getAllSauces = (req, res, next) => {
   // Recherche de sauce déjà existante
   Sauce.find()
-  // Async si sauces existes res.status(200) affichage json contenant la ou les sauces
+    // Async si sauces existes res.status(200) affichage json contenant la ou les sauces
     .then((sauces) => res.status(200).json(sauces))
-  // Sinon recherche erreur retour 404 affichage json contenant objet error
+    // Sinon recherche erreur retour 404 affichage json contenant objet error
     .catch((error) => res.status(404).json({ error }));
 };
 
@@ -16,11 +16,11 @@ exports.createSauce = (req, res, next) => {
   const sauceObject = JSON.parse(req.body.sauce);
   // Créer une nouvelle instance de "Sauce"
   const sauce = new Sauce({
-  // J'utilise le spread operateur qui revient à décomposer le req.body de sauce 
-  // name : sauceObject.name
-  // manufacturer : sauceObject.manufacturer
-  // etc...
-      ...sauceObject,
+    // J'utilise le spread operateur qui revient à décomposer le req.body de sauce
+    // name : sauceObject.name
+    // manufacturer : sauceObject.manufacturer
+    // etc...
+    ...sauceObject,
     // Dans le protocol http sur le serveur création d'un objet qui va être inséré dans la base de données contenant le nom del'image
     imageUrl: `${req.protocol}://${req.get("host")}/images/${
       req.file.filename
@@ -29,7 +29,7 @@ exports.createSauce = (req, res, next) => {
     likes: 0,
     // Dislikes par défaut = 0
     dislikes: 0,
-// Identifiant des utilisateurs qui ont likés 
+    // Identifiant des utilisateurs qui ont likés
     usersLiked: [" "],
     // Identifiant des utilisateurs qui ont dislikés
     usersdisLiked: [" "],
@@ -46,27 +46,31 @@ exports.createSauce = (req, res, next) => {
 exports.getOneSauce = (req, res, next) => {
   // Récupération de l'identifiant dans le corps de la requête
   Sauce.findOne({ _id: req.params.id })
-  // Async retour status 200 avec un json contenant la sauce
+    // Async retour status 200 avec un json contenant la sauce
     .then((sauce) => res.status(200).json(sauce))
     // Recherche erreur, retour status 404 avec un objet contenant error
     .catch((error) => res.status(404).json({ error }));
 };
 exports.deleteSauce = (req, res, next) => {
-  // Récupération de l'identifiant dans le corps de la requête
+  // Récupération de l'identifiant dans l'url de la requête
   Sauce.findOne({ _id: req.params.id })
-  // Async
+    // Async
     .then((sauce) => {
-      // Affectation à filename du nom de l'image
-      const filename = sauce.imageUrl.split("/images/")[1];
-      // Utilisation de la propriété unlink de fileSysteme 
-      fs.unlink(`images/${filename}`, () => {
-        // Suppression de l'image dans le dossier
-        Sauce.deleteOne({ _id: req.params.id })
-        // Async, si tt ok (res.status(200) avec msg Sauce supprimée
-          .then(res.status(200).json({ message: "Sauce supprimée" }))
-          // Sinon recherche erreur retour status 400 avec objet erreur
-          .catch((error) => res.status(400).json({ error }));
-      });
+      if (req.auth.userId !== sauce.userId) {
+        res.status(403).json({ message: "  unauthorized request ! " });
+      } else {
+        // Affectation à filename du nom de l'image
+        const filename = sauce.imageUrl.split("/images/")[1];
+        // Utilisation de la propriété unlink de fileSysteme
+        fs.unlink(`images/${filename}`, () => {
+          // Suppression de l'image dans le dossier
+          Sauce.deleteOne({ _id: req.params.id })
+            // Async, si tt ok (res.status(200) avec msg Sauce supprimée
+            .then(res.status(200).json({ message: "Sauce supprimée" }))
+            // Sinon recherche erreur retour status 400 avec objet erreur
+            .catch((error) => res.status(400).json({ error }));
+        });
+      }
     })
     // Recherche erreur et retour status 500 avec un objet contenant l'erreur
     .catch((error) => res.status(500).json({ error }));
@@ -75,28 +79,32 @@ exports.deleteSauce = (req, res, next) => {
 exports.updateSauce = (req, res, next) => {
   // Affectation du corps de la requête à sauceObject
   const sauceObject = req.file
-  // If ("Terner")
-  // Si il existe un req.file
-    ? {
-      // Parse du req.body.sauce
+    ? // If ("Ternaire")
+      // Si il existe un req.file
+      {
+        // Parse du req.body.sauce
         ...JSON.parse(req.body.sauce),
         // changement de la valeur de imageUrl et du filename
         imageUrl: `${req.protocol}://${req.get("host")}/images/${
           req.file.filename
         }`,
       }
-      // Sinon seulement req.body car il n'y a pas de file
-    : { ...req.body };
-  // Mise à jour de la sauce
-  Sauce.updateOne(
-    // Update de la sauce en envoyant l'id en param
-    { _id: req.params.id },
-    { ...sauceObject, _id: req.params.id }
-  )
-  // Si ok res.status(200) avec objet contenant message "Sauce modifiée"
-    .then(res.status(200).json({ message: "Sauce modifiée" }))
-    // Recherche erreur res.status(400) avec json contenant message erreur
-    .catch((error) => res.status(400).json({ error }));
+    : // Sinon seulement req.body car il n'y a pas de file
+      { ...req.body };
+  if (req.auth.userId !== sauceObject.userId) {
+    res.status(403).json({ message: "  unauthorized request ! " });
+  } else {
+    // Mise à jour de la sauce
+    Sauce.updateOne(
+      // Update de la sauce en envoyant l'id en param
+      { _id: req.params.id },
+      { ...sauceObject, _id: req.params.id }
+    )
+      // Si ok res.status(200) avec objet contenant message "Sauce modifiée"
+      .then(res.status(200).json({ message: "Sauce modifiée" }))
+      // Recherche erreur res.status(400) avec json contenant message erreur
+      .catch((error) => res.status(400).json({ error }));
+  }
 };
 
 exports.likeDislikeSauce = (req, res, next) => {
@@ -107,9 +115,9 @@ exports.likeDislikeSauce = (req, res, next) => {
   // userId = id dans le req.params
   let sauceId = req.params.id;
 
-  // Test de différents cas 
+  // Test de différents cas
   switch (like) {
-    // Cas 1 
+    // Cas 1
     case 1:
       // Méthode pour mise à jour de la sauce
       Sauce.updateOne(
@@ -118,7 +126,7 @@ exports.likeDislikeSauce = (req, res, next) => {
         // Push dans le tableau usersLiked et incrémentation du like
         { $push: { usersLiked: userId }, $inc: { likes: +1 } }
       )
-      // Async réponse status 200, objet contenant message `J'aime`
+        // Async réponse status 200, objet contenant message `J'aime`
         .then(() => res.status(200).json({ message: `J'aime` }))
         // Recherche erreur res.status(400) avec objet contenant l'error
         .catch((error) => res.status(400).json({ error }));
@@ -128,7 +136,7 @@ exports.likeDislikeSauce = (req, res, next) => {
     case 0:
       // Recherche d'une sauce par l'identifiant
       Sauce.findOne({ _id: sauceId })
-      // Async recherche de la présence de l'userId et de la sauce dans le tableau like
+        // Async recherche de la présence de l'userId et de la sauce dans le tableau like
         .then((sauce) => {
           if (sauce.usersLiked.includes(userId)) {
             // Méthode pour mise à jour de la sauce
@@ -140,7 +148,7 @@ exports.likeDislikeSauce = (req, res, next) => {
             )
               // Async réponse status 200, objet contenant message `Neutre`
               .then(() => res.status(200).json({ message: `Neutre` }))
-               // Recherche erreur res.status(400) avec objet contenant l'error
+              // Recherche erreur res.status(400) avec objet contenant l'error
               .catch((error) => res.status(400).json({ error }));
           }
           // Recherche de la présence de l'userId et de la sauce dans le tableau dislike
@@ -174,10 +182,10 @@ exports.likeDislikeSauce = (req, res, next) => {
           // Async réponse status 200, objet contenant message `Je n'aime pas`
           res.status(200).json({ message: `Je n'aime pas` });
         })
-         // Recherche erreur res.status(400) avec objet contenant l'error
+        // Recherche erreur res.status(400) avec objet contenant l'error
         .catch((error) => res.status(400).json({ error }));
       break;
-        // Comportement en cas de problème 
+    // Comportement en cas de problème
     default:
       // Log erreur
       console.log(error);
